fourth commit

+++ src/Passenger.java
@@ -2,14 +2,14 @@ public abstract class Passenger {
     private String name;
     private  int age;
     public  String address;
-    private int idCard;
+    private long idCard;
     private long contact;
     public String getAddress(){return  address;}
     public void setAddress(String address){this.address=address;}
     public String getName(){return name;}
     public void setName(String name){this.name=name;}
-    public int getIdCard(){return  idCard;}
-    public void setIdCard(int idCard){this.idCard=idCard;}
+    public long getIdCard(){return  idCard;}
+    public void setIdCard(long idCard){this.idCard=idCard;}
     public long getContact(){return  contact;}
     public void setContact(long contact){this.contact=contact;}
     public int getAge(){return  age;}
